requirements.txt
# Flask==3.0.0
# python-dotenv==1.0.1
# requests==2.31.0
# mysql-connector-python==8.2.0
# google-generativeai==0.4.0
# python-telegram-bot==21.0
# babel==2.12.1

fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.1
requests==2.31.0
mysql-connector-python==8.2.0
google-generativeai==0.4.0
python-telegram-bot==21.0
babel==2.12.1